# swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
      "/lessons": {
        "get": {
          "operationId": "LessonsController_getLessons",
          "parameters": [
            {
              "name": "topic",
              "required": false,
              "in": "query",
              "description": "Search lesson using topic",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "date",
              "required": false,
              "in": "query",
              "description": "Search lesson using date",
              "schema": {
                "format": "date-time",
                "type": "string"
              }
            },
            {
              "name": "courseId",
              "required": false,
              "in": "query",
              "description": "Search lesson using curse id",
              "schema": {
                "type": "number"
              }
            },
            {
              "name": "areaId",
              "required": false,
              "in": "query",
              "description": "Search lesson using area id",
              "schema": {
                "type": "number"
              }
            },
            {
              "name": "instituteId",
              "required": false,
              "in": "query",
              "description": "Institute id of user",
              "schema": {
                "type": "number"
              }
            },
            {
              "name": "exist",
              "required": false,
              "in": "query",
              "description": "Search group if exist or not",
              "schema": {
                "type": "boolean"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Lessons Routes"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "post": {
          "operationId": "LessonsController_createLesson",
          "parameters": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLessonDto"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": ""
            }
          },
          "tags": [
            "Lessons Routes"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
      "/lessons/{id}": {
        "get": {
          "operationId": "LessonsController_getLessonResult",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Lessons Routes"
          ]
        },
        "patch": {
          "operationId": "LessonsController_updateLesson",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "number"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateLessonDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Lessons Routes"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "delete": {
          "operationId": "LessonsController_deleteLesson",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Lessons Routes"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
      "/lessons/{id}/answers": {
        "get": {
          "operationId": "LessonsController_getAnswersByLesson",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Lessons Routes"
          ]
        }
      },
      "/lessons/{id}/questions": {
        "get": {
          "operationId": "LessonsController_getQuestionsByLesson",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Lessons Routes"
          ]
        }
      },
...
      "/users": {
        "get": {
          "operationId": "UsersController_getUsers",
          "parameters": [
            {
              "name": "name",
              "required": false,
              "in": "query",
              "description": "Search user user by name",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "lastName",
              "required": false,
              "in": "query",
              "description": "Search user user by last name",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "email",
              "required": false,
              "in": "query",
              "description": "Search user user by email",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "code",
              "required": false,
              "in": "query",
              "description": "Search user user by code",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "instituteId",
              "required": false,
              "in": "query",
              "description": "Institute id of user",
              "schema": {
                "type": "number"
              }
            },
            {
              "name": "exist",
              "required": false,
              "in": "query",
              "description": "Search the user if exist or not",
              "schema": {
                "type": "boolean"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Users Routes"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "post": {
          "operationId": "UsersController_createUser",
          "parameters": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserDto"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": ""
            }
          },
          "tags": [
            "Users Routes"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
      "/users/{id}": {
        "get": {
          "operationId": "UsersController_getUser",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Users Routes"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "patch": {
          "operationId": "UsersController_updateUser",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "number"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Users Routes"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "delete": {
          "operationId": "UsersController_deleteUser",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Users Routes"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      }
    },
    "info": {
      "title": "Pandora",
      "description": "Api to make lessons more fun",
      "version": "2.0",
      "contact": {}
    },
...
    "servers": [],
    "components": {
      "securitySchemes": {
        "bearer": {
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "type": "http"
        }
      },
      "schemas": {
...
        "CreateLessonDto": {
          "type": "object",
          "properties": {
            "topic": {
              "type": "string",
              "description": "Topic of lesson"
            },
            "date": {
              "format": "date-time",
              "type": "string",
              "description": "Date of lesson"
            },
            "courseId": {
              "type": "number",
              "description": "Course id of lesson"
            },
            "areaId": {
              "type": "number",
              "description": "Area id of lesson"
            },
            "instituteId": {
              "type": "number",
              "description": "Institute id of user"
            },
            "exist": {
              "type": "boolean",
              "description": "Indicates if the lesson is active or \"deleted\""
            }
          },
          "required": [
            "topic",
            "date",
            "courseId",
            "areaId",
            "instituteId",
            "exist"
          ]
        },
        "UpdateLessonDto": {
          "type": "object",
          "properties": {
            "topic": {
              "type": "string",
              "description": "Topic of lesson"
            },
            "date": {
              "format": "date-time",
              "type": "string",
              "description": "Date of lesson"
            },
            "courseId": {
              "type": "number",
              "description": "Course id of lesson"
            },
            "areaId": {
              "type": "number",
              "description": "Area id of lesson"
            },
            "instituteId": {
              "type": "number",
              "description": "Institute id of user"
            },
            "exist": {
              "type": "boolean",
              "description": "Indicates if the lesson is active or \"deleted\""
            }
          },
          "required": [
            "topic",
            "date",
            "courseId",
            "areaId",
            "instituteId",
            "exist"
          ]
        },
...
        "CreateUserDto": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "description": "Name of user"
            },
            "lastName": {
              "type": "string",
              "description": "Last name of user"
            },
            "email": {
              "type": "string",
              "description": "Email of user"
            },
            "code": {
              "type": "string",
              "description": "Unique user code"
            },
            "instituteId": {
              "type": "number",
              "description": "Institute id of user"
            },
            "password": {
              "type": "string",
              "description": "Password of user"
            },
            "exist": {
              "type": "boolean",
              "description": "Indicates if the question is active or \"deleted\""
            }
          },
          "required": [
            "name",
            "lastName",
            "email",
            "code",
            "instituteId",
            "password",
            "exist"
          ]
        },
        "UpdateUserDto": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "description": "Name of user"
            },
            "lastName": {
              "type": "string",
              "description": "Last name of user"
            },
            "email": {
              "type": "string",
              "description": "Email of user"
            },
            "code": {
              "type": "string",
              "description": "Unique user code"
            },
            "rol": {
              "type": "string",
              "description": "Rol of user"
            },
            "instituteId": {
              "type": "number",
              "description": "Institute id of user"
            },
            "password": {
              "type": "string",
              "description": "Password of user"
            },
            "exist": {
              "type": "boolean",
              "description": "Indicates if the question is active or \"deleted\""
            }
          },
          "required": [
            "name",
            "lastName",
            "email",
            "code",
            "rol",
            "instituteId",
            "password",
            "exist"
          ]
        },
...
      }
    }
  }